requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
buildozer==0.29
ccsm==0.9.11.3
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
compizconfig-python==0.9.11.3
Cython==0.21.2
debtagshw==0.1
defer==1.0.6
dirspec==13.10
dnspython==1.11.1
docutils==0.11
duplicity==0.6.23
future==0.14.3
gnome-app-install===0.4.7-nmu1ubuntu3
gpxpy==0.9.8
html5lib==0.999
httplib2==0.8
Kivy==1.9.0
Kivy-Garden==0.1.1
lockfile==0.8
lxml==3.3.3
Mako==0.9.1
MarkupSafe==0.18
mercurial==2.8.2
oauthlib==0.6.1
oneconf==0.3.7
PAM==0.4.2
pdfshuffler==0.6.0
pexpect==3.3
Pillow==2.3.0
piston-mini-client==0.7.5
plyer==1.2.4
py==1.4.27
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
Pygments==1.6
pygobject==3.12.0
pyjnius==1.3.0
pyOpenSSL==0.13
pyPdf==1.13
pyserial==2.6
pysmbc==1.0.14.1
pytest==2.7.1
python-apt===0.9.3.5ubuntu1
python-dateutil==1.5
python-debian===0.1.21-nmu2ubuntu2
pyxdg==0.25
reportlab==3.0
requests==2.2.1
roman==2.0.0
six==1.5.2
software-center-aptd-plugins==0.0.0
system-service==0.1.6
Twisted-Core==13.2.0
Twisted-Web==13.2.0
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==1.11.4
wheel==0.24.0
xdiagnose===3.6.3build2
zope.interface==4.0.5
